Move to utility package
