+++ Floods/src/mesh/Mesh.java
@@ -3,6 +3,8 @@ package mesh;
 import com.jme3.math.Vector2f;
 import com.jme3.math.Vector3f;
 
+import utility.Location;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.BlockingDeque;
