+++ Floods/src/utility/ServiceResponse.java
@@ -1,4 +1,4 @@
-package services;
+package utility;
 
 public class ServiceResponse extends Message {
